refactored _translate to _prepareTranslate and created new _translate
